--- src/components/extension/index.ts
@@ -1,2 +1,0 @@
-export { ShareIntentRoot } from './ShareIntentRoot';
-export { IOSShareIntentRoot } from './IOSShareIntentRoot';
